requirements.txt
aniso8601==8.0.0
astroid==2.4.2
attrs==20.3.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
dnspython==1.16.0
Flask==1.1.2
Flask-RESTful==0.3.8
gunicorn==20.0.4
idna==2.10
iniconfig==1.1.1
ipaddress==1.0.23
ipdata==3.2
ipwhois==1.1.0
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.16.0
lazy-object-proxy==1.4.3
lxml==4.6.3
MarkupSafe==1.1.1
mccabe==0.6.1
numpy==1.20.1
packaging==20.9
pandas==1.2.3
peewee==3.13.3
pluggy==0.13.1
psycopg2-binary==2.8.6
py==1.9.0
pylint==2.5.3
pyparsing==2.4.7
pyshark==0.4.2.11
pytest==6.2.3
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
requests-file==1.5.1
scikit-learn==0.24.1
scipy==1.6.1
six==1.15.0
threadpoolctl==2.1.0
tld==0.12.2
tldextract==2.2.2
toml==0.10.1
typed-ast==1.4.1
urllib3==1.25.9
websockets==8.1
Werkzeug==1.0.1
wrapt==1.12.1
